Enables instrumented tests on CI
Configures GitHub Actions to run instrumented tests on an Android emulator.

This change adds steps to enable KVM for faster emulator performance,
sets up an Android emulator with API level 34 and Pixel 8a profile,
executes the instrumented tests using Gradle, and uploads the coverage
report to Codacy.

It also adds a task to generate a JaCoCo report for the instrumented tests.